Fix Windows compilation warnings
git-svn-id: http://svn.cern.ch/guest/lhcb/Boole/trunk@41017 4525493e-7705-40b1-a816-d608a930855b

Former-commit-id: 3a5d1c742e42cb6a4e9b17b4202f0d34c2c0ec29

+++ Boole/Muon/MuonAlgs/src/MuonCardiacChannelOutput.cpp
@@ -23,7 +23,7 @@ void  MuonCardiacChannelOutput::addPhyChannel(MuonPhysicalChannelOutput* value)
     //	  m_Hits.push_back(po);
     if( po->getMCMuonHistory().isHitFiring() ) {
       MuonCardiacTraceBack pi(po);      
-      float time= (*po).hitArrivalTime();      
+      float time= (float)((*po).hitArrivalTime());      
 //      std::cout<<chID()<<" "<<time<<std::endl;
       m_timeList.push_back(std::pair<float,MuonCardiacTraceBack>(time,pi));      
     }
